+++ example/lib/counter_simple_cube/counter_simple_cube.dart
@@ -1,0 +1,9 @@
+import 'package:cubes/cubes.dart';
+
+class CounterSimpleCube extends SimpleCube {
+  final count = 0.obsValue;
+
+  void increment() {
+    count.modify((value) => value + 1);
+  }
+}