+++ lib/src/util/state_mixin.dart
@@ -17,28 +17,32 @@ mixin StateMixin<T extends StatefulWidget> on State<T> {
 /// Mixin to user Cube in StatefulWidget
 mixin CubeStateMixin<T extends StatefulWidget, C extends Cube> on State<T> {
   /// Cube that will be used
-  C? cube;
+  C? _cube;
+
+  set cube(C cube) => _cube = cube;
+  C get cube => _cube!;
 
   /// Initial data used in `cube.onReady`
   Object? get initData => null;
 
   @override
   void initState() {
-    cube = cube ?? Cubes.getDependency();
-    cube?.addOnActionListener(_innerOnAction);
+    _cube = _cube ?? Cubes.getDependency();
+    _cube?.addOnActionListener(_innerOnAction);
     WidgetsBinding.instance?.addPostFrameCallback(_ready);
     super.initState();
   }
 
   @override
   void dispose() {
-    cube?.dispose();
+    _cube?.dispose();
     removeOnActionListener();
     super.dispose();
   }
 
   /// Remove action listeners.
-  void removeOnActionListener() => cube?.removeOnActionListener(_innerOnAction);
+  void removeOnActionListener() =>
+      _cube?.removeOnActionListener(_innerOnAction);
 
   void _innerOnAction(C cube, CubeAction action) => onAction(action);
 
@@ -47,6 +51,6 @@ mixin CubeStateMixin<T extends StatefulWidget, C extends Cube> on State<T> {
 
   void _ready(_) {
     var data = initData ?? ModalRoute.of(context)?.settings.arguments;
-    cube?.onReady(data);
+    _cube?.onReady(data);
   }
 }