+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 1.1.0
+
+- Add `SimpleCube`
+- Update `GetIt`
+- Add `Cubes.registerDependencyAsync`
+- Add `Cubes.getDependencyAsync`
+- improvements in `CubeStateMixin`
+
 # 1.0.2
 
 - Update get_it